Refactor GearCard component import paths for consistency and organization
