+++ components/gear/gear-card.tsx
